requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
wikipedia
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langserve
fastapi
langchain_groq
langchain_core
sse_starlette
uvicorn